move CUDA support to a package extension
