+++ ext/CUDAExt/utils.jl
@@ -1,0 +1,1 @@
+rng_from_array(::CuArray) = CUDA.default_rng()
