--- src/cuda/cuda.jl
@@ -1,11 +1,0 @@
-module CUDAint
-
-using ..CUDA
-
-import ..Flux: Flux
-using ChainRulesCore
-import NNlib, NNlibCUDA
-
-include("cudnn.jl")
-
-end
